Use type int32 for the thresholds of nominal conditions.

+++ python/subprojects/common/mlrl/common/cython/rule_model.pyx
@@ -27,8 +27,8 @@ cdef class ConjunctiveBody:
                   const uint32[::1] numerical_gr_indices, const float32[::1] numerical_gr_thresholds,
                   const uint32[::1] ordinal_leq_indices, const float32[::1] ordinal_leq_thresholds,
                   const uint32[::1] ordinal_gr_indices, const float32[::1] ordinal_gr_thresholds,
-                  const uint32[::1] nominal_eq_indices, const float32[::1] nominal_eq_thresholds,
-                  const uint32[::1] nominal_neq_indices, const float32[::1] nominal_neq_thresholds):
+                  const uint32[::1] nominal_eq_indices, const int32[::1] nominal_eq_thresholds,
+                  const uint32[::1] nominal_neq_indices, const int32[::1] nominal_neq_thresholds):
         """
         :param numerical_leq_indices:       A contiguous array of type `uint32`, shape `(num_numerical_leq_conditions)`,
                                             that stores the feature indices of the numerical conditions that use the <=
@@ -57,13 +57,13 @@ cdef class ConjunctiveBody:
         :param nominal_eq_indices:          A contiguous array of type `uint32`, shape `(num_nominal_eq_conditions)`,
                                             that stores the feature indices of the nominal conditions that use the ==
                                             operator or None, if no such conditions are available
-        :param nominal_eq_thresholds:       A contiguous array of type `float32`, shape `(num_nominal_eq_conditions)`
-                                            that stores the thresholds of the nominal conditions that use the ==
-                                            operator or None, if no such conditions are available
+        :param nominal_eq_thresholds:       A contiguous array of type `int32`, shape `(num_nominal_eq_conditions)` that
+                                            stores the thresholds of the nominal conditions that use the == operator or
+                                            None, if no such conditions are available
         :param nominal_neq_indices:         A contiguous array of type `uint32`, shape `(num_nominal_neq_conditions)`,
                                             that stores the feature indices of the nominal conditions that use the !=
                                             operator or None, if no such conditions are available
-        :param nominal_neq_thresholds:      A contiguous array of type `float32`, shape `(num_nominal_neq_conditions)`
+        :param nominal_neq_thresholds:      A contiguous array of type `int32`, shape `(num_nominal_neq_conditions)`
                                             that stores the thresholds of the nominal conditions that use the !=
                                             operator or None, if no such conditions are available
         """
@@ -248,14 +248,14 @@ cdef class RuleList(RuleModel):
         cdef uint32 num_nominal_eq = body.getNumNominalEq()
         cdef const uint32[::1] nominal_eq_indices = \
             <uint32[:num_nominal_eq]>body.nominal_eq_indices_cbegin() if num_nominal_eq > 0 else None
-        cdef const float32[::1] nominal_eq_thresholds = \
-            <float32[:num_nominal_eq]>body.nominal_eq_thresholds_cbegin() if num_nominal_eq > 0 else None
+        cdef const int32[::1] nominal_eq_thresholds = \
+            <int32[:num_nominal_eq]>body.nominal_eq_thresholds_cbegin() if num_nominal_eq > 0 else None
 
         cdef uint32 num_nominal_neq = body.getNumNominalNeq()
         cdef const uint32[::1] nominal_neq_indices = \
             <uint32[:num_nominal_neq]>body.nominal_neq_indices_cbegin() if num_nominal_neq > 0 else None
-        cdef const float32[::1] nominal_neq_thresholds = \
-            <float32[:num_nominal_neq]>body.nominal_neq_thresholds_cbegin() if num_nominal_neq > 0 else None
+        cdef const int32[::1] nominal_neq_thresholds = \
+            <int32[:num_nominal_neq]>body.nominal_neq_thresholds_cbegin() if num_nominal_neq > 0 else None
 
         self.visitor.visit_conjunctive_body(
             ConjunctiveBody.__new__(ConjunctiveBody, numerical_leq_indices, numerical_leq_thresholds,
@@ -304,11 +304,11 @@ cdef class RuleList(RuleModel):
                 if num_ordinal_gr > 0 else None,
             np.asarray(<uint32[:num_ordinal_gr]>body.ordinal_gr_indices_cbegin()) \
                 if num_ordinal_gr > 0 else None,
-            np.asarray(<float32[:num_nominal_eq]>body.nominal_eq_thresholds_cbegin()) \
+            np.asarray(<int32[:num_nominal_eq]>body.nominal_eq_thresholds_cbegin()) \
                 if num_nominal_eq > 0 else None,
             np.asarray(<uint32[:num_nominal_eq]>body.nominal_eq_indices_cbegin()) \
                 if num_nominal_eq > 0 else None,
-            np.asarray(<float32[:num_nominal_neq]>body.nominal_neq_thresholds_cbegin()) \
+            np.asarray(<int32[:num_nominal_neq]>body.nominal_neq_thresholds_cbegin()) \
                 if num_nominal_neq > 0 else None,
             np.asarray(<uint32[:num_nominal_neq]>body.nominal_neq_indices_cbegin()) \
                 if num_nominal_neq > 0 else None,
@@ -346,9 +346,9 @@ cdef class RuleList(RuleModel):
         cdef const uint32[::1] ordinal_leq_indices = body_state[5]
         cdef const float32[::1] ordinal_gr_thresholds = body_state[6]
         cdef const uint32[::1] ordinal_gr_indices = body_state[7]
-        cdef const float32[::1] nominal_eq_thresholds = body_state[8]
+        cdef const int32[::1] nominal_eq_thresholds = body_state[8]
         cdef const uint32[::1] nominal_eq_indices = body_state[9]
-        cdef const float32[::1] nominal_neq_thresholds = body_state[10]
+        cdef const int32[::1] nominal_neq_thresholds = body_state[10]
         cdef const uint32[::1] nominal_neq_indices = body_state[11]
         cdef uint32 num_numerical_leq = numerical_leq_thresholds.shape[0] if numerical_leq_thresholds is not None else 0
         cdef uint32 num_numerical_gr = numerical_gr_thresholds.shape[0] if numerical_gr_thresholds is not None else 0
